requirements.txt
requests
beautifulsoup4
psycopg2-binary
PyMuPDF
python-dotenv
sendgrid
langchain
openai
gradio
pandas